Subject: Moved NextUIProvider to page root
Body:
In this commit, we optimized the application's hierarchy by moving the NextUIProvider wrapper to the root of each individual page instead of having it wrap the entire application. This was done to reduce unnecessary rerenders at the app level, improving performance.

The changes also included a streamlined import to use a single import line from @nextui-org/react instead of separate lines from @nextui-org/{component-name} for simplicity and improved readability.

The SignUp Welcome page has been removed as it became redundant with these updates.

Remember to check each usage of NextUI components carefully to ensure that it’s properly wrapped by NextUIProvider.

Issues: N/A