Subject: Add route to 'Experiences' on 'Next' click from Bio signup page
Body:
In this revision, we've integrated routing functionality into the 'Bio' signup page of our application. Now, when users complete filling in their bio information and click on the 'Next' button, they will be seamlessly redirected to the 'Experiences' signup page. This change aims to improve the navigation flow during the signup process, thus enhancing user experience by reducing the necessity for manual intervention or navigation.

No issues or bugs have been identified in relation to this update.