feat: Add migration for creating Gender table and update User model
The User model has been updated to include a new field called 'genderId'. This field is of type integer and is used as a foreign key to reference the 'id' column of the Gender table. Additionally, a migration file has been added to create the Gender table with columns for 'id' and 'name'. Indexes and constraints have also been added to ensure data integrity.