feat: Add language and location table modifications to support multiple languages and locations
This commit adds the necessary changes to the database schema and signup form to support multiple languages and locations. It modifies the "Language" and "Location" tables by dropping the "userId" column and adding a new "locationId" column to the "User" table. It also creates a new table called "_UserLanguages" to establish a many-to-many relationship between users and languages. These changes improve accessibility and user experience by enabling users to select multiple languages and locations during the signup process.