```text refactor: Remove "title" column from "Experience" table
This commit removes the "title" column from the "Experience" table in the database schema. The column is dropped using an ALTER TABLE statement in the migration file. The removal of the "title" column is necessary to align with the recent code changes and improve the data model. This change ensures that all data in the "title" column will be lost.