Refactor authentication and navigation UI components
The key focus of these changes is to improve user navigation and the authentication process across a number of pages and components, streamlining the experience and codebase.

The 'Home' component code is refactored, replacing previous navigation links with buttons to deliver a more intuitive interface. The check for whether a user is authenticated now determines the display of a 'Login' button, rather than multiple elements as previously.

Within the 'MenuNavbar', there are significant changes in data handling. The user information no longer comes from the component's local state but is now passed as a property, reducing dependence on side effects and making the component more predictable. A visual update is also evident as the profile link now directs users to their socials page rather than their profile.

Logging during the authentication process is adjusted on two pages for a cleaner console output during debugging.

Lastly, the 'Navbar' component is updated to accommodate these changes, with the 'user' object and session information now being passed to the 'MenuNavbar'.

Together these enhancements foster a cleaner, more maintainable code base, while simultaneously improving the user's navigational experience.