Subject: Enforce unique user emails and enhance user deletion
Body: Made a crucial modification to the Prisma schema to enforce uniquess of user emails, mitigating potential data collision issues. Enhanced the user-deletion process to now include deleting associated accounts and collections. Tokens linked with these collections are now also ensured to be deleted to maintain data integrity.

Additionally, a workaround was implemented in user sign in to set emails as null due to an existing bug in Next Auth throwing errors when emails are not null.

Issue References: Next Auth generated issue